use resizeBoundingBox, move absolute helpers to resize helpers
